# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Pazuri",
        "description": "Pazuri"
    },
    "paths": {
...
        "/auth/change-password": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "ChangePassword",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ChangePassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.ChangePassword"
            }
        },
...
        "/friend/send-request": {
            "post": {
                "x-controller-name": "FriendController",
                "x-operation-name": "SendRequest",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SendRequest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "opponent_user_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "FriendController.SendRequest"
            }
        },
        "/friend/requested-list": {
            "get": {
                "x-controller-name": "FriendController",
                "x-operation-name": "RequestedList",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RequestedList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "page",
                    "in": "query",
                    "schema": {
                        "type": "number"
                    }
                }],
                "operationId": "FriendController.RequestedList"
            }
        },
        "/friend/requested-action": {
            "post": {
                "x-controller-name": "FriendController",
                "x-operation-name": "RequestedAction",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RequestedAction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "request_id": {
                                        "type": "string"
                                    },
                                    "action": {
                                        "type": "string",
                                        "enum": ["accept", "reject"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "FriendController.RequestedAction"
            }
        },
        "/friend/friends-list": {
            "get": {
                "x-controller-name": "FriendController",
                "x-operation-name": "FriendsList",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "FriendsList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "page",
                    "in": "query",
                    "schema": {
                        "type": "number"
                    }
                }],
                "operationId": "FriendController.FriendsList"
            }
        },
        "/friend/remove-friend": {
            "post": {
                "x-controller-name": "FriendController",
                "x-operation-name": "RemoveFriend",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RemoveFriend",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "opponent_user_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "FriendController.RemoveFriend"
            }
        },
        "/friend/location-list": {
            "get": {
                "x-controller-name": "FriendController",
                "x-operation-name": "LocationList",
                "tags": ["FriendController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LocationList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "page",
                    "in": "query",
                    "schema": {
                        "type": "number"
                    }
                }],
                "operationId": "FriendController.LocationList"
            }
        },
        "/user/user-list": {
            "get": {
                "x-controller-name": "UserController",
                "x-operation-name": "UserList",
                "tags": ["UserController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UserList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "UserController.UserList"
            }
        },
        "/user/details": {
            "get": {
                "x-controller-name": "UserController",
                "x-operation-name": "Details",
                "tags": ["UserController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "opponent_user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "UserController.Details"
            }
        }
    },
...
    "servers": [{
        "url": "http://localhost:1550/api"
    }]
}